+++ Target/Variable/Environment.d.ts
@@ -1,9 +1,12 @@
-export declare const string: typeof import("zod").string;
+export declare const string: typeof import("zod").string, coerce: typeof import("node_modules/zod/v4/classic/coerce.d.cts");
 declare const _default: import("zod").ZodObject<{
+    Batch: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodCoercedNumber<unknown>>>;
+    Delay: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodCoercedNumber<unknown>>>;
     Email: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
     ID: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
     Key: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
-    Token: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
+    Logger: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodCoercedNumber<unknown>>>;
     Project: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
+    Token: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
 }, import("zod/v4/core").$strip>;
 export default _default;